Rename aocs directory
